% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_targets.R
\name{add_mixed_targets_total_relative}
\alias{add_mixed_targets_total_relative}
\title{Add mixed total targets (relative to baseline)}
\usage{
add_mixed_targets_total_relative(x, targets, overwrite = FALSE, label = NULL)
}
\description{
Adds relative mixed-total targets as proportions of baseline totals per feature.
}
