% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_objectives.R
\name{add_objective_min_intervention_fragmentation}
\alias{add_objective_min_intervention_fragmentation}
\title{add objective: minimize intervention fragmentation}
\usage{
add_objective_min_intervention_fragmentation(
  x,
  relation_name = "boundary",
  weight_multiplier = 1,
  alias = NULL
)
}
\arguments{
\item{x}{A \code{\link{Data}} object created with \code{\link[=inputData]{inputData()}} / \code{\link[=inputDataSpatial]{inputDataSpatial()}}.}

\item{relation_name}{character. Name of the stored relation in \code{x$data$spatial_relations}.}

\item{weight_multiplier}{numeric. Multiplies all relation weights (BLM-like). Default 1.}

\item{alias}{character or NULL. Optional unique identifier to register this objective
as an atomic objective for multi-objective workflows.}
}
\value{
Updated \code{\link{Data}} object.
}
\description{
Stores an objective specification to later build a model that minimizes \emph{intervention-level}
fragmentation over a given spatial relation.

If \code{alias} is provided, the objective is also registered in \code{x$data$objectives}
as an atomic objective for multi-objective workflows.
}
