% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_targets.R
\name{add_recovery_targets_relative}
\alias{add_recovery_targets_relative}
\title{Add recovery targets (relative)}
\usage{
add_recovery_targets_relative(
  x,
  targets,
  relative_basis = c("potential", "baseline"),
  overwrite = FALSE,
  label = NULL
)
}
\arguments{
\item{x}{A \code{Data} object.}

\item{targets}{Target specification (proportions).}

\item{relative_basis}{"potential" (default) or "baseline".}

\item{overwrite}{Logical. If TRUE, replace existing recovery targets for the same features.}

\item{label}{Optional label.}
}
\value{
Updated \code{Data} object.
}
\description{
Adds relative recovery targets specified as proportions in \link{0,1} of a chosen basis.
By default, basis is the potential maximum improvement per feature from dist_benefit.
}
