% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOproblem-class.R
\name{as_mo_problem}
\alias{as_mo_problem}
\title{Create a multi-objective problem wrapper}
\usage{
as_mo_problem(x)
}
\arguments{
\item{x}{A \code{Data} object created by \code{prioriactions::inputData()} (or a compatible object).}
}
\value{
An object of class \code{MOProblem}.
}
\description{
Create a multi-objective problem wrapper
}
\keyword{internal}
