% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-class.R
\name{data-class}
\alias{data-class}
\alias{Data}
\title{Data class}
\value{
No return value.
}
\description{
This class is used to represent data of the instances of the corresponding
multi-action planning problem. It includes several methods for retrieving the information
of the instance (such as the spatial allocation of threats and species, the cost
of management actions or the structure of the spatial connectivity across
the area where the planning is carried out. This class is created using the
\code{\link[=inputData]{inputData()}} function.
}
\section{Fields}{
 \describe{
\item{data}{\code{list} object containing data.}
}
}

\section{Methods}{
 \describe{
\item{getActionsAmount():}{\code{integer}. Number of possible actions.}
\item{getData(\code{character} name):}{\code{\link[=data.frame]{data.frame()}}. Object stored in the \code{data} field with name \code{name}.}
\item{getFeatureAmount():}{\code{integer}. Number of features.}
\item{getFeatureNames():}{\code{character}. Names of features.}
\item{getMonitoringCosts():}{\code{numeric} \code{\link[=vector]{vector()}}. Cost of monitoring each planning unit.}
\item{getPlanningUnitsAmount():}{\code{integer}. Number of planning units.}
\item{getActionCosts():}{\code{numeric} \code{\link[=vector]{vector()}}. Cost of actions each planning unit and threat.}
\item{getThreatNames():}{\code{character}. Names of threats.}
\item{getThreatsAmount():}{\code{integer}. Number of threats.}
\item{print():}{Print basic information of the data instance and (if present) the model state.}
\item{show():}{Call print method.}
}
}

