% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solution.R
\name{get_actions}
\alias{get_actions}
\title{Get action results from a Solution}
\usage{
get_actions(x, only_selected = FALSE)
}
\arguments{
\item{x}{A \linkS4class{solution} object returned by \code{\link[=solve]{solve()}}.}

\item{only_selected}{Logical. If \code{TRUE}, return only selected actions
(\code{selected == 1}). Default \code{FALSE}.}
}
\value{
A \code{data.frame} with action allocation information and a \code{selected} column.
}
\description{
Extracts the action-allocation table from a \linkS4class{solution} object returned by \code{\link[=solve]{solve()}}.
The returned table includes a \code{selected} column indicating whether each feasible
\code{(pu, action)} pair is selected in the solution.
}
\examples{
\dontrun{
sol <- solve(hola)
act_tbl <- get_actions(sol)
act_sel <- get_actions(sol, only_selected = TRUE)
}
}
