% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_solution.R
\name{get_pu}
\alias{get_pu}
\title{Get planning unit results from a Solution}
\usage{
get_pu(x, only_selected = FALSE)
}
\arguments{
\item{x}{A \linkS4class{solution} object returned by \code{\link[=solve]{solve()}}.}

\item{only_selected}{Logical. If \code{TRUE}, return only selected planning units.
Default \code{FALSE}.}
}
\value{
A \code{data.frame} with planning-unit information and a \code{selected} column.
}
\description{
Extracts the planning-unit table from a \linkS4class{solution} object returned by \code{\link[=solve]{solve()}}.
The returned table includes a \code{selected} column indicating whether each planning
unit is selected in the solution.
}
\examples{
\dontrun{
sol <- solve(hola)
pu_tbl <- get_pu(sol)
pu_sel <- get_pu(sol, only_selected = TRUE)
}
}
