% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_objectives.R
\name{set_objective_min_cost}
\alias{set_objective_min_cost}
\title{Set objective: minimize costs}
\usage{
set_objective_min_cost(x, include_pu_cost = TRUE, include_action_cost = TRUE)

set_objective_min_cost(x, include_pu_cost = TRUE, include_action_cost = TRUE)
}
\arguments{
\item{x}{A \code{\link{Data}} object created with \code{\link[=inputData]{inputData()}} / \code{\link[=inputDataSpatial]{inputDataSpatial()}}.}

\item{include_pu_cost}{logical. If \code{TRUE}, include planning-unit costs in the objective.}

\item{include_action_cost}{logical. If \code{TRUE}, include action costs in the objective.}
}
\value{
Updated \code{\link{Data}} object.

Updated \code{\link{Data}} object.
}
\description{
Stores the objective specification inside the \code{\link{Data}} object so it can be
materialized later when the optimization model is built (typically from \code{solve()}).

This objective minimizes the total cost associated with planning units and/or actions,
depending on the flags provided.

Stores the objective specification inside the \code{\link{Data}} object so it can be
materialized later when the optimization model is built (typically from \code{solve()}).

This objective minimizes the total cost associated with planning units and/or actions,
depending on the flags provided.
}
\details{
This function does \strong{not} build or solve the optimization model.
It only updates \code{x$data$model_args}:
\describe{
\item{\code{model_type}}{\code{"minimizeCosts"}}
\item{\code{objective_id}}{\code{"min_cost"}}
\item{\code{objective_args}}{a list with \code{include_pu_cost} and \code{include_action_cost}}
}

If you call another objective setter afterwards, it will overwrite the previous objective.

This function does \strong{not} build or solve the optimization model.
It only updates \code{x$data$model_args}:
\describe{
\item{\code{model_type}}{\code{"minimizeCosts"}}
\item{\code{objective_id}}{\code{"min_cost"}}
\item{\code{objective_args}}{a list with \code{include_pu_cost} and \code{include_action_cost}}
}

If you call another objective setter afterwards, it will overwrite the previous objective.
}
\examples{
\dontrun{
x <- inputDataSpatial(pu = pu_sf, cost = "cost", features = feat_sf, pu_id_col = "id") \%>\%
  add_actions(actions_df) \%>\%
  set_objective_min_cost()

# Later:
# sol <- solve(x)
}

\dontrun{
x <- inputDataSpatial(pu = pu_sf, cost = "cost", features = feat_sf, pu_id_col = "id") \%>\%
  add_actions(actions_df) \%>\%
  set_objective_min_cost()

# Later:
# sol <- solve(x)
}

}
