% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show}
\alias{show}
\alias{show,Problem-method}
\alias{show,Solution-method}
\alias{show,SolutionSet-method}
\title{Show}
\usage{
\S4method{show}{Problem}(x)

\S4method{show}{SolutionSet}(x)
}
\arguments{
\item{x}{Any object.}
}
\description{
Display information about an object.
}
\seealso{
\code{\link[methods:show]{methods::show()}}.
}
\keyword{internal}
