% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sim.R
\docType{data}
\name{sim_dist_features}
\alias{sim_dist_features}
\title{Simulated feature distribution}
\format{
A data frame linking planning units and features with an \code{amount} column.
}
\usage{
data(sim_dist_features)
}
\description{
Example distribution of feature amounts across planning units.
}
\keyword{datasets}
