% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sim.R
\docType{data}
\name{sim_features}
\alias{sim_features}
\title{Simulated features}
\format{
A data frame with feature identifiers and names.
}
\usage{
data(sim_features)
}
\description{
Example feature table for package examples and tests.
}
\keyword{datasets}
