% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_area_constraints.R
\name{add_area_min_constraint}
\alias{add_area_min_constraint}
\title{Add minimum selected area constraint}
\usage{
add_area_min_constraint(
  x,
  area_min,
  area_col = NULL,
  area_unit = c("m2", "ha", "km2"),
  name = "area_min"
)
}
\arguments{
\item{x}{A \linkS4class{data} object.}

\item{area_min}{Numeric. Minimum area to select (in \code{area_unit}).}

\item{area_col}{Character. Column in \code{x$data$pu} containing area (optional).}

\item{area_unit}{Character. Units for \code{area_min}: "m2", "ha", or "km2".}

\item{name}{Optional constraint name.}
}
\value{
Updated \linkS4class{data} object.
}
\description{
Adds \eqn{\sum_i area_i w_i \ge A_{min}} using only PU selection vars \code{w_i}.
}
