% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_targets.R
\name{add_conservation_targets_absolute}
\alias{add_conservation_targets_absolute}
\title{Add conservation targets (absolute)}
\usage{
add_conservation_targets_absolute(x, targets, overwrite = FALSE, label = NULL)
}
\arguments{
\item{x}{A \code{Data} object.}

\item{targets}{See \strong{Targets format} in Details.}

\item{overwrite}{If TRUE, replace existing conservation targets for the same features.}

\item{label}{Optional label stored with the targets.}
}
\value{
Updated \code{Data} object.
}
\description{
Adds absolute conservation targets for each feature.
}
