% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_targets.R
\name{add_conservation_targets_relative}
\alias{add_conservation_targets_relative}
\title{Add conservation targets (relative to baseline)}
\usage{
add_conservation_targets_relative(x, targets, overwrite = FALSE, label = NULL)
}
\arguments{
\item{x}{A \code{Data} object.}

\item{targets}{Target specification (proportions).}

\item{overwrite}{Logical. If TRUE, replace existing conservation targets for the same features.}

\item{label}{Optional label.}
}
\value{
Updated \code{Data} object.
}
\description{
Adds relative conservation targets specified as proportions in \link{0,1} of the
total baseline representation of each feature in the study area.
}
