% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_targets.R
\name{add_recovery_targets_absolute}
\alias{add_recovery_targets_absolute}
\title{Add recovery targets (absolute)}
\usage{
add_recovery_targets_absolute(x, targets, overwrite = FALSE, label = NULL)
}
\arguments{
\item{x}{A \code{Data} object.}

\item{targets}{Target specification.}

\item{overwrite}{Logical. If TRUE, replace existing recovery targets for the same features.}

\item{label}{Optional label.}
}
\value{
Updated \code{Data} object.
}
\description{
Adds absolute recovery targets (absolute amount of recovery gain required).
}
