% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Solution}
\alias{plot.Solution}
\title{Plot a Solution (if spatial geometry is available)}
\usage{
\method{plot}{Solution}(
  x,
  what = c("pu", "actions"),
  only_selected = FALSE,
  action = NULL,
  ...,
  include_conservation = TRUE,
  conservation_label = "conservation",
  base_alpha = 0.1,
  selected_alpha = 0.9,
  base_fill = "grey60",
  base_color = NA,
  selected_color = NA
)
}
\description{
Plot a Solution (if spatial geometry is available)
}
