% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve}
\alias{solve}
\title{Solve optimization model}
\usage{
solve(x, ...)
}
\arguments{
\item{x}{A \code{Data} object created with \code{inputData()} or \code{inputDataSpatial()}.}
}
\value{
A \code{Solution} object.
}
\description{
Solves a model defined by a \code{Data} object. Solver configuration is read from
\code{x$data$solve_args} (typically set via \code{set_solver()} / \code{set_solver_*()}).
}
